feat(server): add dedicated class for listening server
Commit for wrapping the sock_fd for the listening server. When
initializing the server, it throws a runtime_error and therefore exits
the program.

Also removed the close(socket_fd) call in main.cpp, because this is
handled by the socket destructor.

Move 'Listening to Port' into the server + also add the port which can
be individualized.

GH-27

Signed-off-by: Markus Kurz <m.kurz92@gmail.com>